рефакторинг кода
