+++ package.json
@@ -13,7 +13,7 @@
   },
   "dependencies": {
 
-    "sass": "^1.86.0"
+    "sass": "^1.86.0",
 
     "classnames": "^2.5.1",
     "react": "^19.1.0",
