+++ src/app.tsx
@@ -1,4 +1,5 @@
 import { GameTimer } from './components/game-timer/game-timer'
+import { GameArea } from './pages/game-area'
 import { GameSetup } from './pages/game-setup'
 import { TestTimer } from './pages/test-timer/TestTimer'
 
@@ -11,7 +12,8 @@ export const App = () => {
         time="01:22"
         movesTaken={4}
       /> */}
-      <TestTimer />
+      {/* <TestTimer /> */}
+      <GameArea/>
     </div>
   )
 }
