+++ src/components/common/player-card/index.ts
