+++ src/components/common/player-card/player-card.module.scss
