+++ src/components/common/player-card/player-card.tsx
@@ -1,13 +1,11 @@
-import style from './player-card.module.scss'
+import { IPlayerCardProps } from '../../../core/types/for-common'
 
-interface IPlayerCardProps {
-  playerName: string
-  playerPoint: number
-}
+import styles from './player-card.module.scss'
+
 
 export const PlayerCard = ({ playerName, playerPoint }: IPlayerCardProps) => {
   return (
-    <div className={style.playerCard}>
+    <div className={styles.playerCard}>
       <h4>{playerName}</h4>
       <h3>{playerPoint}</h3>
     </div>
