--- src/components/game-area-main/game-area-main.tsx
@@ -1,14 +1,0 @@
-import style from './game-area-main.module.scss'
-import { AreaCard } from '../common/area-card/index'
-
-interface IGameAreaMainProps {}
-
-export const GameAreaMain = ({}: IGameAreaMainProps) => {
-  return (
-    <div className={style.gameAreaMain}>
-      {Array.from({ length: 36 }, (_, index) => (
-        <AreaCard key={index} />
-      ))}
-    </div>
-  )
-}
