+++ src/components/game-area/game-area-footer/game-area-footer.module.scss
