+++ src/components/game-area/game-area-footer/index.ts
@@ -1,0 +1,1 @@
+export { GameAreaFooter } from './game-area-footer'
