+++ src/components/game-area/game-area-header/game-area-header.module.scss
