+++ src/components/game-area/game-area-header/index.ts
