+++ src/components/game-area/game-area-main/game-area-main.module.scss
