+++ src/components/game-area/game-area-main/index.ts
