+++ src/components/game-area/index.ts
@@ -1,0 +1,3 @@
+export { GameAreaMain } from './game-area-main'
+export { GameAreaHeader } from './game-area-header'
+export { GameAreaFooter } from './game-area-footer'
