--- src/components/game-footer/index.ts
@@ -1,1 +1,0 @@
-export { GameFooter } from './game-footer'
