+++ src/core/types/for-common.ts
@@ -3,3 +3,15 @@ export interface IActionButtonProps {
   children: React.ReactNode
   variant?: 'orange' | 'secondary'
 }
+
+export interface IAreaButtonProps {
+    buttonText: string
+    onClick: () => void
+}
+
+export interface IAreaCardProps {}
+
+export interface IPlayerCardProps {
+    playerName: string
+    playerPoint: number
+}
