+++ src/core/types/for-game-area.ts
@@ -1,0 +1,6 @@
+export interface IGameAreaMainProps {}
+
+export interface IGameAreaHeaderProps {}
+
+export interface IGameAreaFooterProps {}
+
