+++ src/pages/game-area/game-area.module.scss
