+++ src/pages/game-area/index.ts
